Use precision 16

+++ db/migrate/20200417042243_create_votes.rb
@@ -2,7 +2,7 @@ class CreateVotes < ActiveRecord::Migration[6.0]
   def change
     create_table :votes do |t|
       t.string :contact_no
-      t.decimal :amount, precision: 10, scale: 2
+      t.decimal :amount, precision: 16, scale: 2
       t.references :charity, foreign_key: true, index: true, null: false
       t.string :status
